Last part of this update of the iced version badge
Elysium doesn't seem to use Iced anymore:
https://github.com/kalmari246/elysium/commit/98f50890f605917eda47b448107bc589d7070a21

Linkage use the 'advanced-text' branch, so I put 0.10 as that was the
closest release to that code:
https://github.com/linkage-rs/linkage/commit/65964661ecc6de6ca6b1223e03951505c34d6621

Neothesia uses a recent revision from the master branch, so I set it as
master (also fixed it's description).

From my understanding the Cosmic stuff uses libcosmic, which by it's
turn keeps an up-to-date fork of Iced, so I set their badge to 'master'.